CategoryService: usage check before delete
Check if any products using this category exist before deletion

+++ category-service/src/main/java/de/hkamicroservices/categoryservice/CategoryController.java
@@ -2,15 +2,21 @@ package de.hkamicroservices.categoryservice;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.reactive.function.client.WebClientResponseException;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+import java.util.Objects;
 
 
 @RestController()
 @RequestMapping(path="/categories")
 public class CategoryController {
 
+    private final String productServiceEndpoint = Objects.equals(System.getenv("PRODUCT_ENDPOINT"), "") ? System.getenv("PRODUCT_ENDPOINT") : "localhost";
     private final CategoryRepository categoryRepository;
 
     @Autowired
@@ -37,8 +43,26 @@ public class CategoryController {
     }
 
     @DeleteMapping(path="/{id}")
-    public void deleteCategory(@PathVariable Long id){
+    public ResponseEntity<?> deleteCategory(@PathVariable Long id){
         if(!categoryRepository.existsById(id)) throw new CategoryNotFoundException();
+        if(checkProductsExist(id)) return ResponseEntity.badRequest().body("Cannot delete category if products using this category exist");
+
         categoryRepository.deleteById(id);
+        return ResponseEntity.ok("");
+    }
+
+    private boolean checkProductsExist(Long categoryId){
+        WebClient client = WebClient.create("http://" + productServiceEndpoint + ":8081/products/");
+        try {
+            var response = client.get().uri("?categoryId="+String.valueOf(categoryId)).accept(MediaType.APPLICATION_JSON).retrieve().bodyToMono(ProductDto[].class).block();
+            assert response != null;
+            return response.length >0;
+            //return response.length > 0;
+        } catch (WebClientResponseException wcre) {
+            return false;
+        }catch (Exception e){
+            System.out.println(e);
+            return true;
+        }
     }
 }
